Add files via upload
Matt's attempt at scoring for the RF that Adhiraj built. Tidyverse and rlang were not working for me. You will have to uncomment the lines that say uncomment. There is one near the top and then a large block that is indented in the middle. Also be sure to save the model in the end of the previous file using:
#Save Model#
save(Mean, Mode, RF, file="project_RF.RData")